BT tools
*Added stop for parallel child nodes
*Added decision node (BT version of rule sheet)
*Added scoring system for decision node
*Changed some names
*Simple leaf - a node that either returns success of fail within the
same frame that it is executed. Points to a function with no arguments
and a boolean return value.
*Complex leaf - can do everything a simple node does, but can run over
more than one frame. In other works, can return IsRunning and Stopped.
REturns a IEnumerator thing and takes a Stopper object as an argument.
Don't worry too much how it works, I'll provide some examples.
*Condition leaf - meant to be used with the Decision node. Pair this
node with whatever kind of node (can be a node or a leaf) to build
decision nodes. Meant to be evaluated within a single frame, returns a
boolean and takes a Scorer object as argument (again, examples will
follow)